Move tests for reinitialising partitions
